// src/shared/types/response.ts
export type ApiResponse<T = null> = {
  isSuccess: boolean;
  message: string;
  messageCode: ErrorCode | null;
  errors?: Record<string, string[]> | null;
  data: T | null;
};

export enum HttpStatusCode {
  // 2xx — Success
  OK = 200,
  CREATED = 201,
  ACCEPTED = 202,
  NO_CONTENT = 204,

  // 4xx — Client errors
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  PAYMENT_REQUIRED = 402,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  PAYLOAD_TOO_LARGE = 413,
  UNPROCESSABLE_ENTITY = 422,
  TOO_MANY_REQUESTS = 429,

  // 5xx — Server errors
  INTERNAL_SERVER_ERROR = 500,
  NOT_IMPLEMENTED = 501,
  SERVICE_UNAVAILABLE = 503,
}

export enum ErrorCode {
  /* AUTH — CREDENTIALS & ACCESS */
  AUTH_INVALID_CREDENTIALS = "AUTH_INVALID_CREDENTIALS",
  AUTH_UNAUTHORIZED = "AUTH_UNAUTHORIZED",

  /* AUTH — TOKEN (JWT / SESSION) */
  AUTH_TOKEN_EXPIRED = "AUTH_TOKEN_EXPIRED",
  AUTH_SESSION_REVOKED = "AUTH_SESSION_REVOKED",

  /*  VALIDATION  */
  VALIDATION_FAILED = "VALIDATION_FAILED",

  /*  SYSTEM  */
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  TOO_MANY_REQUESTS = "TO_MANY_REQUESTS",
}